disabled old etl tests and verified everything is running e2e
